+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix unit test failure caused by moto 5 upgrade
 - Fix pagination bug when listing glue databases and tables
 - Fix _create_session_config isolation to prevent overrides between sessions
+- Fix tmp table location for the default file_format
 
 ## v1.7.2
 - Fix the issue that removes double quote unexpectedly